+++ test/e2e/replica_set_tls_upgrade/replica_set_tls_upgrade_test.go
@@ -33,7 +33,7 @@ func TestReplicaSetTLSUpgrade(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	if err := setup.CreateTLSResources(mdb.Namespace, ctx); err != nil {
+	if err := setup.CreateTLSResources(mdb.Namespace, ctx, setup.CertKeyPair); err != nil {
 		t.Fatalf("Failed to set up TLS resources: %s", err)
 	}